Move `h.views.api_moderation` -> `h.views.api.moderation`
